Decrease lag when showing profiling data in client
